POC: Split into k8s_sandbox_core + Inspect layer
Add k8s_sandbox_core package containing all K8s/Helm/Pod functionality
with zero inspect_ai dependencies. The existing k8s_sandbox package
becomes a thin re-export layer on top of core, adding Inspect-specific
tracing/logging. Both packages ship in one distribution.

This enables non-Inspect consumers (e.g. a standalone sandboxes service)
to use K8s sandbox primitives (Release, Pod, ExecResult) without pulling
in inspect_ai.

Key changes:
- New k8s_sandbox_core with standalone ExecResult, concurrency, limits,
  and logging (no-op trace_action, logger.debug for traces)
- k8s_sandbox modules become re-export shims from core
- inspect-ai moved from required to optional dependency ([inspect] extra)
- Test mock.patch targets updated to k8s_sandbox_core.X (necessary
  because code now executes in the core module namespace)
