+++ pyproject.toml
@@ -6,14 +6,20 @@ authors = [{name = "UK AI Security Institute"}]
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
-  "inspect-ai>=0.3.161",
   "kubernetes>=35.0.0",
   "jsonschema>=4.23.0",
   "tenacity>=8.0.0",
+  "shortuuid>=1.0.0",
+  "pyyaml>=6.0.0",
+  "rich",
 ]
 
 [project.optional-dependencies]
+inspect = [
+  "inspect-ai>=0.3.161",
+]
 dev = [
+  "inspect-ai>=0.3.161",
   "kubernetes-stubs-elephant-fork>=35.0.0",
   "mypy>=1.9.0",
   "pre-commit>=3.6.2",
@@ -39,7 +45,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/k8s_sandbox"]
+packages = ["src/k8s_sandbox", "src/k8s_sandbox_core"]
 
 
 [tool.ruff.lint]
