+++ src/k8s_sandbox/__init__.py
@@ -1,6 +1,6 @@
 """Package for a Kubernetes sandbox environment provider for Inspect AI."""
 
-from k8s_sandbox._pod import GetReturncodeError, PodError
+from k8s_sandbox_core._pod import GetReturncodeError, PodError
 from k8s_sandbox._sandbox_environment import (
     K8sError,
     K8sSandboxEnvironment,
