+++ src/k8s_sandbox_core/__init__.py
@@ -1,0 +1,30 @@
+"""Core K8s/Helm/Pod functionality without inspect_ai dependency."""
+
+from k8s_sandbox_core._exec_result import ExecResult
+from k8s_sandbox_core._helm import (
+    DEFAULT_CHART,
+    Release,
+    StaticValuesSource,
+    ValuesSource,
+    get_all_release_names,
+    uninstall,
+)
+from k8s_sandbox_core._limits import OutputLimitExceededError, SandboxLimits
+from k8s_sandbox_core._manager import HelmReleaseManager
+from k8s_sandbox_core._pod import GetReturncodeError, Pod, PodError
+
+__all__ = [
+    "DEFAULT_CHART",
+    "ExecResult",
+    "GetReturncodeError",
+    "HelmReleaseManager",
+    "OutputLimitExceededError",
+    "Pod",
+    "PodError",
+    "Release",
+    "SandboxLimits",
+    "StaticValuesSource",
+    "ValuesSource",
+    "get_all_release_names",
+    "uninstall",
+]
