+++ src/k8s_sandbox_core/_exec_result.py
@@ -1,0 +1,16 @@
+"""Lightweight ExecResult dataclass matching inspect_ai's interface."""
+
+from dataclasses import dataclass
+from typing import Generic, TypeVar
+
+T = TypeVar("T", str, bytes)
+
+
+@dataclass
+class ExecResult(Generic[T]):
+    """The result of executing a command."""
+
+    success: bool
+    returncode: int
+    stdout: T
+    stderr: T
