+++ src/k8s_sandbox_core/_pod/error.py
@@ -1,0 +1,30 @@
+from typing import Any
+
+from k8s_sandbox_core._logger import format_log_message
+
+
+class PodError(Exception):
+    """
+    A generic error raised when interacting with a Pod.
+
+    This will typically cause the eval to fail.
+    """
+
+    def __init__(self, message: str, **kwargs: Any) -> None:
+        super().__init__(format_log_message(message, **kwargs))
+
+
+class GetReturncodeError(Exception):
+    """The return code of a Pod operation could not be retrieved."""
+
+    pass
+
+
+class ExecutableNotFoundError(Exception):
+    """The executable could not be found in the container.
+
+    This is **not** raised as a result of user-supplied commands not being found. It is
+    typically raised when /bin/sh or runuser cannot be found in the container.
+    """
+
+    pass
