+++ src/k8s_sandbox_core/resources
@@ -1,0 +1,1 @@
+../k8s_sandbox/resources
