+++ test/k8s_sandbox/pod/test_keepalive.py
@@ -12,7 +12,7 @@ def test_send_keepalive_sends_frames_and_stops():
     ws_client.is_open.return_value = True
     stop = threading.Event()
 
-    with patch("k8s_sandbox._pod.op._KEEPALIVE_INTERVAL_SECONDS", 0.05):
+    with patch("k8s_sandbox_core._pod.op._KEEPALIVE_INTERVAL_SECONDS", 0.05):
         t = threading.Thread(target=_send_keepalive, args=(ws_client, stop))
         t.start()
         sleep(0.2)
@@ -29,7 +29,7 @@ def test_send_keepalive_exits_when_socket_closes():
     ws_client.is_open.return_value = False
     stop = threading.Event()
 
-    with patch("k8s_sandbox._pod.op._KEEPALIVE_INTERVAL_SECONDS", 0.01):
+    with patch("k8s_sandbox_core._pod.op._KEEPALIVE_INTERVAL_SECONDS", 0.01):
         t = threading.Thread(target=_send_keepalive, args=(ws_client, stop))
         t.start()
         t.join(timeout=1)
